Add files via upload
To use this file for upload, you will need to download the MFRC522 library (listed in arduino libraries). I believe that SPI should be a default library!

I can help you change the conditionals and the chip ID numbers based on the code already written. The easiest way to get the card number is to use an app on your phone,